fix: consumption - Allowing supported langs to be ingested on old node engines - Adding required prop to tutorialmodal
